Fix result type and arguments in receive_file() and add file_storage.c in CMakeLists.txt

+++ include/file_storage.h
@@ -1,8 +1,10 @@
 #ifndef FILE_STORAGE_H
 #define FILE_STORAGE_H
 
+#include <unistd.h>
+
 int send_file(int client_socket, const char* path);
 
-int receive_file(int client_socket, const char* filename, unsigned int file_size);
+int receive_file(int client_socket, const char* filename, size_t file_size);
 
 #endif
